+++ abjad/__init__.py
@@ -1,16 +1,3 @@
-# warn on an outdated Python installation
-import distutils.version
-import platform
-if not (
-    distutils.version.LooseVersion('2.7.2') <
-    distutils.version.LooseVersion(platform.python_version())
-    ):
-    print('WARNING: Abjad does not support Python versions less than 2.7.3.')
-    print('WARNING: Upgrade your Python to 2.7.3 or higher.')
-del platform
-del distutils
-
-
 import os
 if 'topleveltools' in os.path.abspath('.'):
     message = 'do not start Abjad from topleveltools directory (inspect.py).'
