+++ abjad/cli/test/base.py
@@ -1,14 +1,21 @@
 import abjad
+import doctest
 import pathlib
 import pytest
+import re
 import shutil
 import sys
+import unittest
+import types
 import uqbar.io
+from io import StringIO
 
 
-class ScorePackageScriptTestCase(abjad.TestCase):
+class ScorePackageScriptTestCase(unittest.TestCase):
     r'''A base test class for score-package scripts.
     '''
+
+    ansi_escape = re.compile(r'\x1b[^m]*m')
 
     test_path = pathlib.Path(__file__).parent
     score_path = test_path.joinpath('test_score')
@@ -21,14 +28,14 @@ class ScorePackageScriptTestCase(abjad.TestCase):
     # ### TEST LIFECYCLE ### #
 
     def setUp(self):
-        super(ScorePackageScriptTestCase, self).setUp()
         if self.score_path.exists():
             shutil.rmtree(self.score_path)
         self.directory_items = set(self.test_path.iterdir())
         sys.path.insert(0, str(self.score_path))
+        self.string_io = StringIO()
 
     def tearDown(self):
-        super(ScorePackageScriptTestCase, self).tearDown()
+        self.string_io.close()
         for path in sorted(self.test_path.iterdir()):
             if path in self.directory_items:
                 continue
@@ -180,3 +187,71 @@ class ScorePackageScriptTestCase(abjad.TestCase):
         parts_path = self.build_path.joinpath('parts.ily')
         with parts_path.open('w') as file_pointer:
             file_pointer.write(pytest.helpers.get_fancy_parts_code())
+
+    # ### PUBLIC METHODS ### #
+
+    def compare_captured_output(self, expected):
+        actual = self.ansi_escape.sub('', self.string_io.getvalue())
+        actual = abjad.String.normalize(actual)
+        expected = abjad.String.normalize(expected)
+        self.compare_strings(expected, actual)
+
+    def compare_file_contents(self, path, expected_contents):
+        expected_contents = abjad.String.normalize(expected_contents)
+        with open(str(path), 'r') as file_pointer:
+            contents = abjad.String.normalize(file_pointer.read())
+        self.compare_strings(expected_contents, contents)
+
+    def compare_lilypond_contents(self, ly_path, expected_contents):
+        expected_contents = abjad.String.normalize(expected_contents)
+        with open(str(ly_path), 'r') as file_pointer:
+            contents = file_pointer.read()
+        if ly_path.suffix == '.ly':
+            contents = contents.splitlines()
+            while 'version' not in contents[0]:
+                contents.pop(0)
+            contents.pop(0)
+            contents = '\n'.join(contents)
+        contents = abjad.String.normalize(contents)
+        self.compare_strings(expected_contents, contents)
+
+    def compare_path_contents(self, path_to_search, expected_files):
+        actual_files = sorted(
+            str(path.relative_to(self.test_path))
+            for path in sorted(path_to_search.glob('**/*.*'))
+            if '__pycache__' not in path.parts and
+            path.suffix != '.pyc'
+            )
+        self.compare_strings(
+            '\n'.join(str(_) for _ in actual_files),
+            '\n'.join(str(_) for _ in expected_files),
+            )
+
+    def compare_strings(self, expected, actual):
+        actual = self.normalize(self.ansi_escape.sub('', actual))
+        expected = self.normalize(self.ansi_escape.sub('', expected))
+        example = types.SimpleNamespace()
+        example.want = expected
+        output_checker = doctest.OutputChecker()
+        flags = (
+            doctest.NORMALIZE_WHITESPACE |
+            doctest.ELLIPSIS |
+            doctest.REPORT_NDIFF
+            )
+        success = output_checker.check_output(expected, actual, flags)
+        if not success:
+            diff = output_checker.output_difference(example, actual, flags)
+            raise Exception(diff)
+
+    def normalize(self, string):
+        return abjad.String.normalize(string)
+
+    def reset_string_io(self):
+        self.string_io.close()
+        self.string_io = StringIO()
+
+    # ### PUBLIC PROPERTIES ### #
+
+    @property
+    def test_path(self):
+        return pathlib.Path(__file__).parent
