+++ abjad/tools/systemtools/__init__.py
@@ -21,7 +21,6 @@ from .Tag import Tag
 from .TemporaryDirectory import TemporaryDirectory
 from .TemporaryDirectoryChange import TemporaryDirectoryChange
 from .TestManager import TestManager
-from .TestCase import TestCase
 from .Timer import Timer
 from .UpdateManager import UpdateManager
 from .WellformednessManager import WellformednessManager
