+++ abjad/tools/test/test_abjad___repr__.py
@@ -11,7 +11,6 @@ ignored_classes = (
     abjad.Path,
     abjad.StorageFormatManager,
     abjad.Tags,
-    abjad.TestCase,
     abjad.book.AbjadDirective,
     abjad.book.CodeBlock,
     abjad.book.CodeOutputProxy,
